refactor(council): extract normalizeCouncillorModels to a pure helper
Move the councillor model-chain normalization into a standalone
utils module with no schema/validation dependency, so CouncilManager
no longer transitively pulls in zod via council-schema. The schema
module re-exports the helper so existing imports are unaffected.

Addresses Greptile P2 on PR #664 (runtime dependency from manager
value-importing the schema module).
